% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Implan hunting spending categories}
\format{A data frame with 46 rows and 4 columns"
\describe{
  \item{type}{"trip" or "equip"}
  \item{item}{food, lodge, etc.}
  \item{category}{implan spending category}
  \item{share}{category share for item}
}}
\usage{
categories
}
\description{
A dataset to relate national survey spending data with implan spending categories
prepared from `system.file("extdata", "implan-categories.xlsx", package = "implan")`
}
\seealso{
Other data: 
\code{\link{sector_scheme536}},
\code{\link{sectors536}},
\code{\link{sectors546}},
\code{\link{sectors_crosswalk}},
\code{\link{spending}}
}
\concept{data}
\keyword{datasets}
