% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{check_implan_sums}
\alias{check_implan_sums}
\title{Check whether the implan Excel sheets match spending by Ind/Comm}
\usage{
check_implan_sums(dat, filename, ..., print_compare = FALSE)
}
\arguments{
\item{dat}{data frame with spending by sector with 4 required columns:
group ("Ind" or "Comm"), sector (numeric), retail ("Yes" or "No"), spend (numeric)}

\item{filename}{file path for output excel file}

\item{...}{Optional grouping variables (unquoted) for separating sheets
across one or more dimensions}

\item{print_compare}{If TRUE, show comparison between the 2 summaries}
}
\description{
This is intended to be called after \code{\link{input}}. It checks sums for
"Ind" and "Comm" separately, based on grouping supplied by the dots argument.
}
\examples{
# see ?input_prep
}
\seealso{
Other functions for implan input: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{input_write}()},
\code{\link{input}()}
}
\concept{functions for implan input}
