% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{check_message_error}
\alias{check_message_error}
\title{Helper function for error message on failed check}
\usage{
check_message_error(dat_not_passed, error_message)
}
\arguments{
\item{dat_not_passed}{data frame to show why there is an error}

\item{error_message}{error message to be printed in front of dat_not_passed}
}
\description{
Helper function for error message on failed check
}
