% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{check_share_sums}
\alias{check_share_sums}
\title{Check whether the share variable sums to 100 percent}
\usage{
check_share_sums(df, sharevar, ...)
}
\arguments{
\item{df}{input data frame}

\item{sharevar}{unquoted name of variable that identifies share}

\item{...}{unquoted variables to use as dimensions (that defines share allocation)}
}
\description{
Check whether the share variable sums to 100 percent
}
\examples{
# see ?sector_update()
}
\seealso{
Other functions to get spending by implan sector: 
\code{\link{sector_update}()}
}
\concept{functions to get spending by implan sector}
