% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{check_spend_sums}
\alias{check_spend_sums}
\title{Check whether reallocated spending retains the correct sum}
\usage{
check_spend_sums(df_old, df_new, spendvar, ...)
}
\arguments{
\item{df_old}{data frame with spending before reallocation}

\item{df_new}{data frame with spending post-reallocation}

\item{spendvar}{unquoted name of spending variable}

\item{...}{unquoted variables to use as grouping dimension}
}
\description{
Check whether reallocated spending retains the correct sum
}
\examples{
# see ?implan_prepare()
}
\seealso{
Other functions to get spending by implan sector: 
\code{\link{check_share_sums}()},
\code{\link{sector_update}()}
}
\concept{functions to get spending by implan sector}
