% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{implan_header}
\alias{implan_header}
\title{Get a header table for Implan import}
\usage{
implan_header(activity_type, activity_name, event_year)
}
\arguments{
\item{activity_type}{either "Industry Change" or "Commodity Change"}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}
}
\description{
This is a convenience function called from \code{\link{implan_prepare}}
}
\examples{
# see ?implan_prepare
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{implan_prepare}}
}
\concept{functions to transfer to/from implan}
