% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{implan_prepare}
\alias{implan_prepare}
\alias{implan_prepare_ind}
\title{Prepare spending by sector for Excel implan import}
\usage{
implan_prepare_ind(dat, activity_name, event_year = 2019)
}
\arguments{
\item{dat}{data frame with spending by sector}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}
}
\description{
This function splits an input data frame into a list with 2 data frames:
(1) a sheet header with activity details used by implan, and (2) the
sector spending with additional columns which implan will calculate
}
\section{Functions}{
\itemize{
\item \code{implan_prepare_ind}: Prepare industry data
}}

\examples{
# get necessary sectoring
data(sector_scheme536, sectors_crosswalk, sectors546)
sector_scheme546 <- sector_update(sector_scheme536, sectors_crosswalk, sectors546)

# calculate total spending by sector
library(dplyr)
data(spending, categories)

spend_category <- spending \%>\%
    left_join(categories, by = c("type", "item")) \%>\%
    mutate(spend = spend * share)
check_spend_sums(spending, spend_category, spend, type, item)

spend_sector <- spend_category \%>\%
    select(-share) \%>\%
    left_join(sector_scheme546, by = "category") \%>\%
    mutate(spend = spend * share)
check_spend_sums(spend_category, spend_sector, spend, type, item, category)

# allocate for implan import (Industry)
ls <- implan_prepare_ind(spend_sector, "huntInd")
ls

# write to an excel file
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{implan_header}()}
}
\concept{functions to transfer to/from implan}
