% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Write spending to Excel for Implan import}
\usage{
input(dat, filename, event_year, ...)
}
\arguments{
\item{dat}{data frame with spending by sector with 4 required columns:
group ("Ind" or "Comm"), sector (numeric), retail ("Yes" or "No"), spend (numeric)}

\item{filename}{file path for output excel file}

\item{event_year}{Activity Year for Implan}

\item{...}{Optional grouping variables (unquoted) for separating sheets
across one or more dimensions}
}
\description{
This writes separate sheets for commodity and industry by wrapping
\code{\link{input_prep}} and \code{\link{xlsx_write_implan}}. The dots
argument allows for an arbitrary number of grouping dimensions (for separate
sheets by dimensions).
}
\examples{
# see ?input_prep()
}
\seealso{
Other functions for implan input: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{xlsx_write_implan}()}
}
\concept{functions for implan input}
