% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_header}
\alias{input_header}
\title{Get a header table for Implan import}
\usage{
input_header(activity_type, activity_name, event_year)
}
\arguments{
\item{activity_type}{either "Industry Change" or "Commodity Change"}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}
}
\description{
This is a convenience function called from \code{\link{input_prep}}
}
\examples{
# see ?input_prep
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_prep}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
