% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_prep}
\alias{input_prep}
\alias{input_prep_ind}
\alias{input_prep_comm}
\title{Prepare spending by sector for Excel implan import}
\usage{
input_prep(dat, activity_name, event_year = 2019, group)

input_prep_ind(dat, activity_name, event_year = 2019)

input_prep_comm(dat, activity_name, event_year = 2019)
}
\arguments{
\item{dat}{data frame with spending by sector with 4 required columns:
group ("Ind" or "Comm"), sector (numeric), retail ("Yes" or "No"), spend (numeric)}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}

\item{group}{Either "Ind" or "Comm"}
}
\description{
This function splits an input data frame into a list with 2 data frames:
(1) a sheet header with activity details used by implan, and (2) the
sector spending with additional columns which implan will calculate. The data
portion is also grouped by sector-retail to ensure the minimum number of rows.
}
\section{Functions}{
\itemize{
\item \code{input_prep_ind}: Prepare industry data

\item \code{input_prep_comm}: Prepare commodity data
}}

\examples{
# get necessary sectoring
data(sector_scheme536, sectors_crosswalk, sectors546)
sector_scheme546 <- sector_update(sector_scheme536, sectors_crosswalk, sectors546)

# calculate total spending by sector
library(dplyr)
data(spending, categories)

spend_category <- spending \%>\%
    left_join(categories, by = c("type", "item")) \%>\%
    mutate(spend = spend * share)
check_spend_sums(spending, spend_category, spend, type, item)

spend_sector <- spend_category \%>\%
    select(-share) \%>\%
    left_join(sector_scheme546, by = "category") \%>\%
    mutate(spend = spend * share)
check_spend_sums(spend_category, spend_sector, spend, type, item, category)

# allocate for implan import (Industry)
comm <- input_prep_comm(spend_sector, "huntComm")
ind <- input_prep_ind(spend_sector, "huntInd")
ind

# write to an excel worksheet
xlsx_write_implan(ind, "tmp.xlsx")
# you'll need to manually save as ".xls" (in Excel) from Implan import
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{output_combine}()},
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
