% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Read and Summarize Implan output}
\usage{
output(dirname)
}
\arguments{
\item{dirname}{directory name that stores files for selected activity}
}
\description{
This wraps \code{\link{output_read_csv}} and \code{\link{output_combine}}. It
also will assume nested directories represent dimensions of interest. For
example, if the top-level directory contains no csv files, it will inspect
sub-directories and append the subdirectory names to a "dimension" variable.
}
\examples{
output_dir <- system.file("extdata", "output", package = "implan")

# we can see the nested struture of these files (i.e., multiple dimensions)
list.files(output_dir, recursive = TRUE)

# pull one dimension of results
dirname <- file.path(output_dir, "region1", "bike")
output(dirname)

# pull all results
df <- output(output_dir)

# parse dimensions
tidyr::separate(df, dimension, c("region", "act"), "/")
}
\seealso{
Other functions for implan output loading: 
\code{\link{output_combine}()},
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()}
}
\concept{functions for implan output loading}
