% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{output_combine}
\alias{output_combine}
\title{Combine csv output for an implan activity into one data frame}
\usage{
output_combine(dat)
}
\arguments{
\item{dat}{list produced by \code{\link{output_read_csv}}}
}
\description{
Puts summary, tax fed, and tax local into a single table
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
