% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{output_format_tax}
\alias{output_format_tax}
\title{A helper function to format Tax Impact csv files}
\usage{
output_format_tax(df, impact_type, tax_type)
}
\arguments{
\item{df}{data frame with tax results}

\item{impact_type}{either "Direct Effect" or "Total Effect"}

\item{tax_type}{either "FedTax" or "LocalTax"}
}
\description{
This simplifies the tax results to just show sum of taxes by direct/total.
Intended to be called from \code{\link{output_combine}}
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{output_combine}()},
\code{\link{output_read_csv}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
