% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{output_match}
\alias{output_match}
\title{A helper function for case insensitive string matching}
\usage{
output_match(string, match)
}
\arguments{
\item{string}{string vector to search}

\item{match}{matching phrase to look for}
}
\description{
A helper function for case insensitive string matching
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{output_combine}()},
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
