% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector_scheme536}
\alias{sector_scheme536}
\title{Implan sectoring scheme (536) for hunting}
\format{A data frame with 140 rows (uniquely identified with category-sector) and 6 columns:
\describe{
  \item{group}{"Ind" or "Comm"}
  \item{category}{spending category (e.g., "Food - Groceries", etc.)}
  \item{sector}{implan sector ID}
  \item{description}{implan sector description}
  \item{share}{share of category to be allocated by sector}
  \item{retail}{"Yes" or "No"}
}}
\usage{
sector_scheme536
}
\description{
A dataset that defines how hunting expenditures will be allocated to implan
sectors, prepared from `system.file("extdata", "implan-sectors536.xlsx", package = "implan")`
}
\seealso{
Other data: 
\code{\link{categories}},
\code{\link{sectors536}},
\code{\link{sectors546}},
\code{\link{sectors_crosswalk}},
\code{\link{spending}}
}
\concept{data}
\keyword{datasets}
