% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{sector_update}
\alias{sector_update}
\title{Update sectoring scheme with crosswalk table}
\usage{
sector_update(
  sectors_old,
  crosswalk,
  description_new,
  id_old = "sector536",
  id_new = "sector546"
)
}
\arguments{
\item{sectors_old}{data frame that holds spending by old sector}

\item{crosswalk}{data frame for converting between new and old sectors}

\item{description_new}{data frame that contains descriptions for new sectors}

\item{id_old}{id_old variable name for the old sector id}

\item{id_new}{variable name for the new sector id}
}
\description{
Update sectoring scheme with crosswalk table
}
\concept{functions to get spending by implan sector}
