% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sectors_crosswalk}
\alias{sectors_crosswalk}
\title{Implan 546/536 crosswalk}
\format{A data frame with 1098 rows and 4 variables:
\describe{
  \item{sector546}{ID for 546}
  \item{sector536}{ID for 536}
  \item{crosswalk_ratio}{for allocation, from variable originally named "CewAvgRatio"}
  \item{group}{"Ind" or "Comm"}
}}
\source{
\url{https://implanhelp.zendesk.com/hc/en-us/articles/360034895094-BEA-Benchmark-The-New-546-Sectoring-Scheme}
}
\usage{
sectors_crosswalk
}
\description{
A dataset for relating different Implan sectoring schemes (546 and 536) prepared from
`system.file("extdata", "Bridge_Implan536ToImplan546.xlsx", package = "implan")`
}
\seealso{
Other data: 
\code{\link{categories}},
\code{\link{sector_scheme536}},
\code{\link{sectors536}},
\code{\link{sectors546}},
\code{\link{spending}}
}
\concept{data}
\keyword{datasets}
