% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spending}
\alias{spending}
\title{Implan total hunting spending}
\format{A data frame with 212 rows and 5 columns"
\describe{
  \item{activity_group}{"oia", "hunt", "fish", "wildlife"}
  \item{act}{"bike", "trail", etc.}
  \item{type}{"trip" or "equip"}
  \item{item}{food, lodge, etc.}
  \item{spend}{total spending}
}}
\usage{
spending
}
\description{
A dataset that defines spending in the format of national survey data prepared
from `system.file("extdata", "spend2019.rds", package = "implan")`
}
\seealso{
Other data: 
\code{\link{category_to_sector536}},
\code{\link{category_to_sector546}},
\code{\link{item_to_category}},
\code{\link{sector536_to_sector546}},
\code{\link{sectors536}},
\code{\link{sectors546}}
}
\concept{data}
\keyword{datasets}
