% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{xlsx_initialize_workbook}
\alias{xlsx_initialize_workbook}
\title{Initialize an Excel Workbook with a README tab}
\usage{
xlsx_initialize_workbook(filename)
}
\arguments{
\item{filename}{path where the Excel workbook will be written}
}
\description{
The README sheet is intended to serve as documentation for the results
stored in the Excel file.
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}},
\code{\link{xlsx_write_implan}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
