% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{xlsx_write_implan}
\alias{xlsx_write_implan}
\title{Write data to a sheet for Excel Implan import}
\usage{
xlsx_write_implan(ls, filename)
}
\arguments{
\item{ls}{list returned from implan_prepare_ind() or implan_prepare_comm()}

\item{filename}{file path for output excel file}
}
\description{
Write data to a sheet for Excel Implan import
}
\examples{
# see ?input_prep()
}
\seealso{
Other functions for implan input: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{input}()}
}
\concept{functions for implan input}
