% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_header}
\alias{input_header}
\title{Get a header table for Implan import}
\usage{
input_header(activity_type, activity_name, event_year)
}
\arguments{
\item{activity_type}{either "Industry Change" or "Commodity Change"}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}
}
\description{
This is a convenience function called from \code{\link{input_prep}}
}
\examples{
# see ?input_prep
}
\seealso{
Other functions for implan input: 
\code{\link{input_prep}()},
\code{\link{input_write}()},
\code{\link{input}()}
}
\concept{functions for implan input}
