% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input_prep}
\alias{input_prep}
\alias{input_prep_ind}
\alias{input_prep_comm}
\title{Prepare spending by sector for Excel implan import}
\usage{
input_prep(dat, activity_name, event_year, group)

input_prep_ind(dat, activity_name, event_year)

input_prep_comm(dat, activity_name, event_year)
}
\arguments{
\item{dat}{data frame with spending by sector with 4 required columns:
group ("Ind" or "Comm"), sector (numeric), retail ("Yes" or "No"), spend (numeric)}

\item{activity_name}{Activity Name used for Implan}

\item{event_year}{Activity Year for Implan}

\item{group}{Either "Ind" or "Comm"}
}
\description{
This function splits an input data frame into a list with 2 data frames:
(1) a sheet header with activity details used by implan, and (2) the
sector spending with additional columns which implan will calculate. The data
portion is also grouped by sector-retail to ensure the minimum number of rows.
}
\section{Functions}{
\itemize{
\item \code{input_prep_ind}: Prepare industry data

\item \code{input_prep_comm}: Prepare commodity data
}}

\examples{
# get necessary sectoring
data(category_to_sector536, sector536_to_sector546, sectors546)
category_to_sector546 <- sector_update(category_to_sector536, sector536_to_sector546, sectors546)

# calculate total spending by sector
library(dplyr)
data(spending, item_to_category)

spend_category <- spending \%>\%
    left_join(item_to_category, by = c("activity_group", "type", "item")) \%>\%
    mutate(spend = spend * share)
check_spend_sums(spending, spend_category, spend, activity_group, type, item)

spend_sector <- spend_category \%>\%
    select(-share) \%>\%
    left_join(category_to_sector546, by = "category") \%>\%
    mutate(spend = spend * share)
check_spend_sums(spend_category, spend_sector, spend, type, item, category)

# allocate for implan import (Industry)
comm <- input_prep_comm(spend_sector, "huntComm", 2019)
ind <- input_prep_ind(spend_sector, "huntInd", 2019)
ind

# write to an excel worksheet
# you'll need to manually save as ".xls" (in Excel) from Implan import
\dontrun{
input_write(ind, "tmp.xlsx")
}

# write sheets by activity
\dontrun{
input(spend_sector, "tmp2.xlsx", 2019, act)
check_implan_sums(spend_sector, "tmp2.xlsx", act)
}
}
\seealso{
Other functions for implan input: 
\code{\link{check_implan_sums}()},
\code{\link{input_header}()},
\code{\link{input_write}()},
\code{\link{input}()}
}
\concept{functions for implan input}
