% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_combine}
\alias{output_combine}
\title{Combine csv output for an implan activity into one data frame}
\usage{
output_combine(dat)
}
\arguments{
\item{dat}{list produced by \code{\link{output_read_csv}}}
}
\description{
Puts summary, tax fed, and tax local into a single table
}
\examples{
# see ?output_read_csv()
}
\seealso{
Other functions for implan output loading: 
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()},
\code{\link{output}()}
}
\concept{functions for implan output loading}
