% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{sector_update}
\alias{sector_update}
\title{Update sectoring scheme with crosswalk table}
\usage{
sector_update(
  scheme_old,
  crosswalk,
  description_new,
  id_old = "sector536",
  id_new = "sector546"
)
}
\arguments{
\item{scheme_old}{data frame that holds old sectoring scheme}

\item{crosswalk}{data frame for converting between new and old sectors}

\item{description_new}{data frame that contains descriptions for new sectors}

\item{id_old}{id_old variable name for the old sector id}

\item{id_new}{variable name for the new sector id}
}
\description{
Update sectoring scheme with crosswalk table
}
\examples{
# load sample data
data(sector_scheme536, sectors_crosswalk, sectors546)

# check the sector allocation for the 536 sector scheme (should be TRUE)
check_share_sums(sector_scheme536, share, category)

# update to 546 sector scheme and check
sector_scheme546 <- sector_update(sector_scheme536, sectors_crosswalk, description_new)
check_share_sums(sector_scheme546, share, category)
}
\seealso{
Other functions to get spending by implan sector: 
\code{\link{check_share_sums}()}
}
\concept{functions to get spending by implan sector}
