% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sectors546}
\alias{sectors546}
\title{Implan 546 sector descriptions}
\format{A data frame with 1092 rows and 3 variables:
\describe{
  \item{sector}{sector ID}
  \item{description}{sector description}
  \item{group}{group: "Ind" or "Comm"}
}}
\source{
\url{https://implanhelp.zendesk.com/hc/en-us/articles/360034895094-BEA-Benchmark-The-New-546-Sectoring-Scheme}
}
\usage{
sectors546
}
\description{
A dataset containing sector IDs and descriptions prepared from
`system.file("extdata", "Implan546 Industries & Commodities.xlsx", package = "implan")`
}
\seealso{
Other data: 
\code{\link{categories}},
\code{\link{sector_scheme536}},
\code{\link{sectors536}},
\code{\link{sectors_crosswalk}},
\code{\link{spending}}
}
\concept{data}
\keyword{datasets}
