% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spending}
\alias{spending}
\title{Implan total hunting spending}
\format{A data frame with 46 rows and 3 columns"
\describe{
  \item{spend_type}{"trip" or "equip"}
  \item{item}{food, lodge, etc.}
  \item{spend}{total spending}
}}
\usage{
spending
}
\description{
A dataset that defines spending in the format of national survey data prepared
from `system.file("extdata", "spend2019.rds", package = "implan")`
}
\seealso{
Other data: 
\code{\link{categories}},
\code{\link{sector_scheme536}},
\code{\link{sectors536}},
\code{\link{sectors546}},
\code{\link{sectors_crosswalk}}
}
\concept{data}
\keyword{datasets}
