% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{xlsx_write_implan}
\alias{xlsx_write_implan}
\title{Write data to a sheet for Excel Implan import}
\usage{
xlsx_write_implan(ls, xls_out)
}
\arguments{
\item{ls}{list returned from implan_prepare_ind() or implan_prepare_comm()}

\item{xls_out}{file path for output excel file}
}
\description{
Write data to a sheet for Excel Implan import
}
\examples{
# see ?input_prep()
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}()},
\code{\link{output_combine}()},
\code{\link{output_format_tax}()},
\code{\link{output_read_csv}()},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_table}()}
}
\concept{functions to transfer to/from implan}
