% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{xlsx_write_table}
\alias{xlsx_write_table}
\title{Write a data frame to an Excel tab}
\usage{
xlsx_write_table(df, tabname, filename)
}
\arguments{
\item{df}{data frame to write to the Excel worksheet}

\item{tabname}{name to use for Excel worksheet}

\item{filename}{path where the Excel workbook will be written}
}
\description{
Requires an existing Excel file, preferably created using
\code{\link{xlsx_initialize_workbook}}. The selected tab will be removed
(if it already exists) and a new tab will be written.
}
\examples{
xlsx_initialize_workbook("tmp.xlsx")
df <- data.frame(a = 1:4, b = c("speak", "friend", "and", "enter"))
xlsx_write_table(df, "moria", "tmp.xlsx")
}
\seealso{
Other functions to transfer to/from implan: 
\code{\link{input_header}()},
\code{\link{input_prep}},
\code{\link{xlsx_initialize_workbook}()},
\code{\link{xlsx_write_implan}()}
}
\concept{functions to transfer to/from implan}
